# Conflicts:
#	.github/workflows/auto_translate.yml
#	.github/workflows/translate_branch_update.yml
#	.gitignore
#	.tgs.yml
#	_maps/map_files/Delta/delta.dmm
#	_maps/map_files/MetaStation/MetaStation.dmm
#	_maps/map_files/RandomRuins/LavaRuins/lavaland_surface_althland_facility.dmm
#	_maps/map_files/RandomRuins/SpaceRuins/oldstation.dmm
#	_maps/map_files/RandomRuins/SpaceRuins/spacebar.dmm
#	_maps/map_files/cerestation/cerestation.dmm
#	_maps/map_files/cyberiad/cyberiad.dmm
#	_maps/map_files220/RandomRuins/LavaRuins/scp_facility.dmm
#	_maps/map_files220/RandomRuins/SpaceRuins/convoy_ambush.dmm
#	_maps/map_files220/RandomRuins/SpaceRuins/infected_ship.dmm
#	_maps/map_files220/RandomRuins/SpaceRuins/mechtransport_new.dmm
#	_maps/map_files220/RandomRuins/SpaceRuins/spacehotel.dmm
#	_maps/map_files220/RandomRuins/SpaceRuins/whiteship.dmm
#	_maps/map_files220/RandomZLevels/beach.dmm
#	_maps/map_files220/RandomZLevels/blackmarketpackers.dmm
#	_maps/map_files220/RandomZLevels/caves.dmm
#	_maps/map_files220/RandomZLevels/gate_lizard.dmm
#	_maps/map_files220/RandomZLevels/wildwest.dmm
#	_maps/map_files220/cyberiad/cyberiad.dmm
#	_maps/map_files220/delta/delta.dmm
#	_maps/map_files220/generic/Admin_Zone.dmm
#	_maps/map_files220/generic/Lavaland.dmm
#	_maps/map_files220/generic/centcomm.dmm
#	code/__DEFINES/misc_defines.dm
#	code/__DEFINES/sound_defines.dm
#	code/__DEFINES/status_effects.dm
#	code/__HELPERS/unsorted.dm
#	code/controllers/subsystem/SSgarbage.dm
#	code/controllers/subsystem/SSticker.dm
#	code/datums/datumvars.dm
#	code/datums/mind.dm
#	code/datums/ruins/lavaland.dm
#	code/datums/spells/cluwne.dm
#	code/game/gamemodes/miniantags/abduction/machinery/experiment.dm
#	code/game/gamemodes/miniantags/demons/slaughter demon/slaughter.dm
#	code/game/gamemodes/miniantags/morph/morph.dm
#	code/game/gamemodes/miniantags/pulsedemon/pulsedemon.dm
#	code/game/gamemodes/miniantags/revenant/revenant.dm
#	code/game/gamemodes/nuclear/nuclear.dm
#	code/game/gamemodes/wizard/artefact.dm
#	code/game/gamemodes/wizard/raginmages.dm
#	code/game/gamemodes/wizard/wizard.dm
#	code/game/machinery/camera/camera.dm
#	code/game/objects/items/devices/scanners.dm
#	code/modules/admin/admin_verbs.dm
#	code/modules/admin/verbs/onlyone.dm
#	code/modules/admin/verbs/striketeam_syndicate.dm
#	code/modules/antagonists/_common/antag_datum.dm
#	code/modules/antagonists/_common/antag_spawner.dm
#	code/modules/antagonists/_common/antag_team.dm
#	code/modules/antagonists/revolutionary/team_revolution.dm
#	code/modules/antagonists/wishgranter/wishgranter_avatar.dm
#	code/modules/client/login_processing/20-load_characters.dm
#	code/modules/client/preference/character.dm
#	code/modules/client/preference/loadout/loadout_racial.dm
#	code/modules/events/traders.dm
#	code/modules/food_and_drinks/food/foods/sandwiches.dm
#	code/modules/mob/living/carbon/human/species/_species.dm
#	code/modules/mob/living/silicon/robot/drone/drone_items.dm
#	code/modules/mob/mob.dm
#	code/modules/mob/mob_misc_procs.dm
#	code/modules/mob/new_player/new_player.dm
#	code/modules/response_team/ert_outfits.dm
#	code/modules/security_levels/security_levels.dm
#	code/modules/telesci/gps.dm
#	config/example/config.toml
#	icons/_nanomaps/Cyberiad220_nanomap_z1.png
#	icons/_nanomaps/Delta220_nanomap_z1.png
#	icons/_nanomaps/MetaStation_nanomap_z1.png
#	icons/mob/clothing/head.dmi
#	icons/mob/clothing/species/drask/head.dmi
#	icons/mob/clothing/species/drask/suit.dmi
#	icons/mob/clothing/species/grey/head.dmi
#	icons/mob/clothing/species/grey/suit.dmi
#	icons/mob/clothing/species/tajaran/mask.dmi
#	icons/mob/clothing/species/vox/head.dmi
#	icons/mob/clothing/species/vox/suit.dmi
#	icons/mob/clothing/suit.dmi
#	icons/mob/human_races/human_skintones/r_human_Coffee.dmi
#	icons/mob/human_races/human_skintones/r_human_Olive.dmi
#	icons/mob/human_races/human_skintones/r_human_Oliverandcompany.dmi
#	icons/mob/human_races/human_skintones/r_human_beige.dmi
#	icons/mob/human_races/human_skintones/r_human_chestnut.dmi
#	icons/mob/human_races/human_skintones/r_human_pale.dmi
#	icons/obj/clothing/hats.dmi
#	icons/obj/clothing/suits.dmi
#	interface/skin.dmf
#	modular_ss220/_modpacks.dm
#	modular_ss220/_signals220/code/signals_mob/signals_mob_living.dm
#	modular_ss220/aesthetics/_aesthetics.dme
#	modular_ss220/aesthetics/boxes/code/boxes.dm
#	modular_ss220/aesthetics/chairs/code/chairs.dm
#	modular_ss220/aesthetics/decals/code/decals.dm
#	modular_ss220/aesthetics/decals/icons/siding.dmi
#	modular_ss220/aesthetics/extinguisher/code/extinguisher.dm
#	modular_ss220/aesthetics/extinguisher/icons/extinguisher.dmi
#	modular_ss220/aesthetics/floors/icons/floors.dmi
#	modular_ss220/aesthetics/racks/code/racks.dm
#	modular_ss220/aesthetics/sheets/code/sheets.dm
#	modular_ss220/aesthetics/sheets/icons/sheets.dmi
#	modular_ss220/aesthetics_sounds/_aesthetics_sounds.dme
#	modular_ss220/balance/_balance.dm
#	modular_ss220/balance/_balance.dme
#	modular_ss220/balance/code/items/weapons.dm
#	modular_ss220/cinematics/_cinematics.dme
#	modular_ss220/cinematics/code/_cinematics.dm
#	modular_ss220/cinematics/code/cinematics/credits.dm
#	modular_ss220/clothing/_clothing.dme
#	modular_ss220/credits/_credits.dme
#	modular_ss220/credits/code/SScredits.dm
#	modular_ss220/credits/icons/logo.dmi
#	modular_ss220/food/code/food.dm
#	modular_ss220/food/icons/food.dmi
#	modular_ss220/gateway/_gateway.dme
#	modular_ss220/jobs/code/cards_ids.dm
#	modular_ss220/maps220/code/Areas/station.dm
#	modular_ss220/maps220/code/directions.dm
#	modular_ss220/maps220/code/effects.dm
#	modular_ss220/maps220/code/floors.dm
#	modular_ss220/maps220/code/mobs.dm
#	modular_ss220/maps220/code/objects.dm
#	modular_ss220/maps220/code/spawners.dm
#	modular_ss220/maps220/icons/decals.dmi
#	modular_ss220/maps220/icons/floors.dmi
#	modular_ss220/mobs/code/simple_animal/hostile/headcrab.dm
#	modular_ss220/modular_ss220.dme
#	modular_ss220/objects/_objects.dm
#	modular_ss220/objects/_objects.dme
#	modular_ss220/objects/code/miscellaneous.dm
#	modular_ss220/objects/code/weapons.dm
#	modular_ss220/objects/icons/ammo.dmi
#	modular_ss220/objects/icons/guns.dmi
#	modular_ss220/security_redalert_accesses/code/security_redalert_accesses.dm
#	modular_ss220/silicons/_silicons.dme
#	modular_ss220/silicons/code/robot_modules.dm
#	modular_ss220/text_to_speech/code/tts_preferences.dm
#	modular_ss220/unique_objects/code/unique_objects.dm
#	modular_ss220/unique_objects/icons/statuelarge.dmi
#	modular_ss220/vending/code/vending.dm
#	modular_ss220/vending/icons/vending.dmi
#	modular_ss220/wire_splicing/wiresplicing.dm
#	tgui/.prettierignore
#	tgui/babel.config.js
#	tgui/package.json
#	tgui/packages/tgui/components/ByondUi.js
#	tgui/packages/tgui/components/NanoMap.js
#	tgui/packages/tgui/interfaces/AgentCard.js
#	tgui/packages/tgui/interfaces/Biogenerator.js
#	tgui/packages/tgui/interfaces/CameraConsole220.js
#	tgui/packages/tgui/interfaces/EmotePanel.js
#	tgui/packages/tgui/interfaces/Jukebox.js
#	tgui/packages/tgui/interfaces/KitchenMachine.js
#	tgui/packages/tgui/interfaces/NTRecruiter.js
#	tgui/packages/tgui/interfaces/SeedExtractor.js
#	tgui/packages/tgui/interfaces/Smartfridge.js
#	tgui/packages/tgui/interfaces/SpecMenu.js
#	tgui/packages/tgui/interfaces/Uplink.js
#	tgui/packages/tgui/public/tgui.bundle.css
#	tgui/packages/tgui/public/tgui.bundle.js
#	tgui/packages/tgui/styles/themes/abductor.scss
#	tgui/yarn.lock
#	tools/translator/converter.py
#	tools/translator/ss220_replacer_linux
#	tools/translator/ss220_replacer_windows.exe

+++ code/modules/mob/living/carbon/carbon.dm
@@ -10,6 +10,8 @@
 		qdel(I)
 	// This clause is here due to items falling off from limb deletion
 	for(var/obj/item in get_all_slots())
+		if(QDELETED(item))
+			continue
 		unEquip(item, silent = TRUE)
 		qdel(item)
 	QDEL_LIST_CONTENTS(internal_organs)
@@ -321,7 +323,7 @@
 			status += " and "
 
 		if(LB.status & ORGAN_BURNT)
-			status += "critically burnt"
+			status += "critically burnt" + (LB.status & ORGAN_SALVED ? ", but salved" : "")
 		else
 			switch(burndamage)
 				if(0.1 to 10)
@@ -619,7 +621,7 @@ GLOBAL_LIST_INIT(ventcrawl_machinery, list(/obj/machinery/atmospherics/unary/ven
 
 /mob/living/carbon/hit_by_thrown_carbon(mob/living/carbon/human/C, datum/thrownthing/throwingdatum, damage, mob_hurt, self_hurt)
 	for(var/obj/item/dualsaber/D in contents)
-		if(HAS_TRAIT(src, TRAIT_WIELDED) && D.force)
+		if(HAS_TRAIT(D, TRAIT_WIELDED) && D.force)
 			visible_message("<span class='danger'>[src] impales [C] with [D], before dropping them on the ground!</span>")
 			C.apply_damage(100, BRUTE, "chest", sharp = TRUE, used_weapon = "Impaled on [D].")
 			C.Stun(2 SECONDS) //Punishment. This could also be used by a traitor to throw someone into a dsword to kill them, but hey, teamwork!
@@ -1345,3 +1347,8 @@ so that different stomachs can handle things in different ways VB*/
 		if(wear_suit.flags_inv & HIDEGLOVES)
 			clean_hands = FALSE
 	..(clean_hands, clean_mask, clean_feet)
+
+/mob/living/carbon/fall_and_crush(turf/target_turf, crush_damage, should_crit, crit_damage_factor, datum/tilt_crit/forced_crit, weaken_time, knockdown_time, ignore_gravity, should_rotate = TRUE, angle)
+	// keep most of what's passed in, but don't change the angle
+	. = ..(target_turf, crush_damage, should_crit, crit_damage_factor, forced_crit, weaken_time, knockdown_time, should_rotate = FALSE, rightable = FALSE)
+	KnockDown(10 SECONDS)
